move functions to cpp
